Refactor and fix edit profile action, and relative service and dao

+++ src/main/java/com/epam/java/rt/lab/dao/Dao.java
@@ -7,7 +7,7 @@ import java.util.List;
  */
 public interface Dao {
 
-    int create(DaoParameter daoParameter) throws DaoException;
+    Long create(DaoParameter daoParameter) throws DaoException;
 
     <T> List<T> read(DaoParameter daoParameter) throws DaoException;
